Step3: Move the base application layer into a pilot base subpackage
Finish the neighborhood moves of the pilot package reorganization with
the hub: _base_app, _gui, _gui_common, and _theme move to
solvcon/pilot/base/.  The package entry __init__.py and the extension
shim _pilot_core.py stay at the package root, so the public
solvcon.pilot surface does not change.

The base subpackage __init__.py binds _gui_common and _base_app before
importing _gui, because _gui pulls in the feature neighborhoods whose
modules reach back into the partially initialized base package for
those two names.  Every other change is the mechanical one-level import
adjustment in the moved modules, the subpackage import sites, the
package root, eight tests, and the setup.py entry.
