+++ setup.py
@@ -76,6 +76,7 @@ def main():
             'solvcon.pilot._euler',
             'solvcon.pilot.agent',
             'solvcon.pilot.airfoil',
+            'solvcon.pilot.base',
             'solvcon.pilot.canvas',
             'solvcon.pilot.onedim',
             'solvcon.pilot.panel',
