+++ solvcon/pilot/__init__.py
@@ -21,7 +21,7 @@ from ._pilot_core import (  # noqa: F401
     RManager,
 )
 if enable:
-    from ._gui import (  # noqa: F401
+    from .base._gui import (  # noqa: F401
         controller,
         launch,
     )
