+++ solvcon/pilot/base/_gui.py
@@ -9,26 +9,26 @@ Graphical-user interface code
 # Use flake8 http://flake8.pycqa.org/en/latest/user/error-codes.html
 
 
-from .. import apputil
-from . import _pilot_core as _pcore
-from . import airfoil
+from ... import apputil
+from .. import _pilot_core as _pcore
+from .. import airfoil
 
 if _pcore.enable:
     from . import _gui_common
-    from .visual import _mesh
-    from .panel import _tree_panel
-    from ._euler import _solution_info
-    from ._euler import _oblique
-    from .onedim import _euler1d
-    from .onedim import _burgers1d
-    from .canvas import _svg_gui
-    from .onedim import _linear_wave
-    from .canvas import _canvas_gui
-    from .canvas import _painter_gui
-    from .panel import _profiling
-    from .agent import _agent_gui
+    from ..visual import _mesh
+    from ..panel import _tree_panel
+    from .._euler import _solution_info
+    from .._euler import _oblique
+    from ..onedim import _euler1d
+    from ..onedim import _burgers1d
+    from ..canvas import _svg_gui
+    from ..onedim import _linear_wave
+    from ..canvas import _canvas_gui
+    from ..canvas import _painter_gui
+    from ..panel import _profiling
+    from ..agent import _agent_gui
     from . import _theme
-    from .panel import _window_manager
+    from ..panel import _window_manager
 
 __all__ = [  # noqa: F822
     'controller',
