+++ solvcon/pilot/canvas/_svg_gui.py
@@ -12,7 +12,7 @@ from PySide6 import QtCore, QtWidgets
 from ... import core
 from ... import apputil
 from ...plot import svg
-from .. import _gui_common
+from ..base import _gui_common
 
 __all__ = [  # noqa: F822
     'SVGFileDialog',
