+++ solvcon/pilot/onedim/_euler1d.py
@@ -4,7 +4,7 @@
 
 from ...onedim import euler1d
 
-from .. import _base_app
+from ..base import _base_app
 
 
 class Euler1DApp(_base_app.OneDimBaseApp):
