+++ solvcon/pilot/visual/_mesh.py
@@ -11,7 +11,7 @@ import os
 from PySide6 import QtCore, QtGui, QtWidgets
 
 from ... import core
-from .. import _gui_common
+from ..base import _gui_common
 
 __all__ = [  # noqa: F822
     'SampleMesh',
