+++ tests/test_pilot_solution_info.py
@@ -13,7 +13,7 @@ from solvcon.multidim.euler import oblique
 
 try:
     from solvcon import pilot
-    from solvcon.pilot import _gui
+    from solvcon.pilot.base import _gui
     from solvcon.pilot._euler import _solution_info
     from PySide6.QtWidgets import QApplication
 except ImportError:
