+++ tests/test_pilot_window_menu.py
@@ -17,7 +17,7 @@ import solvcon
 
 try:
     from solvcon import pilot
-    from solvcon.pilot import _gui
+    from solvcon.pilot.base import _gui
     from solvcon.pilot.onedim import _euler1d
     from solvcon.pilot.panel._window_manager import WindowManager
     from PySide6 import QtWidgets
